+++ serverside/makoto/src/main/kotlin/AttestationService.kt
@@ -3,14 +3,23 @@
 package at.asitplus.attestation
 
 import at.asitplus.attestation.AttestationException
+import at.asitplus.attestation.Makoto.Companion.assertionReader
 import at.asitplus.catchingUnwrapped
 import at.asitplus.signum.indispensable.AndroidKeystoreAttestation
 import at.asitplus.signum.indispensable.Attestation
 import at.asitplus.signum.indispensable.IosHomebrewAttestation
 import at.asitplus.signum.indispensable.toJcaPublicKey
+import ch.veehait.devicecheck.appattest.assertion.Assertion
+import ch.veehait.devicecheck.appattest.assertion.AssertionChallengeValidator
+import ch.veehait.devicecheck.appattest.attestation.ValidatedAttestation
+import ch.veehait.devicecheck.appattest.common.AuthenticatorData
+import ch.veehait.devicecheck.appattest.common.AuthenticatorDataFlag
 import kotlinx.coroutines.runBlocking
 import java.security.PublicKey
 import java.security.cert.X509Certificate
+import java.security.interfaces.ECPublicKey
+import kotlin.experimental.and
+import kotlin.experimental.xor
 import kotlin.time.ExperimentalTime
 import at.asitplus.attestation.AttestationException as AttException
 
@@ -207,22 +216,83 @@ abstract class AttestationService {
         /**
          * Verifies an App Attestation in conjunction with an assertion for some client data.
          *
-         * First, it verifies the app attestation, afterwards it verifies the assertion, checks whether at most [counter] many signatures
-         * have been performed using the key bound to the attestation before signing the assertion and verifies whether the client data
+         * First, it verifies the app attestation, afterwards it verifies the assertion, checks whether at most the assertion was the first thing that was signed
+         * after attesting the key (i.e. signature counter = `0` before signing the assertion), and verifies whether the client data
          * referenced within the assertion matches [referenceClientData]
          *
          * @param attestationObject the AppAttest attestation object to verify
          * @param assertionFromDevice the assertion data created on the device.
          * @param referenceClientData the expected client data to be contained in [assertionFromDevice]
-         * @param counter the highest expected value of the signature counter prior to creating the assertion.
          */
+        fun verifyCombined(
+            attestationObject: ByteArray,
+            assertionFromDevice: ByteArray,
+            referenceClientData: ByteArray,
+            challenge: ByteArray,
+        ): AttestationResult
+
+        @Deprecated(
+            "Misnomer, counter is ignored",
+            level = DeprecationLevel.ERROR,
+            replaceWith = ReplaceWith("verifyCombined(attestationObject, assertionFromDevice, referenceClientData, challenge)")
+        )
         fun verifyAssertion(
             attestationObject: ByteArray,
             assertionFromDevice: ByteArray,
             referenceClientData: ByteArray,
             challenge: ByteArray,
             counter: Long = 0
-        ): AttestationResult
+        ) = verifyCombined(attestationObject, assertionFromDevice, referenceClientData, challenge)
+
+        /**
+         * Verifies a fresh assertion, tied to a previously stored attestation.
+         * This function assumes that `clientDataHash` is the SHA-256 digest of [expectedChallenge].
+         *
+         * **The attestation is not verified again!**
+         * There is no timeliness guarantee of any kind, so manually verify the freshness of the challenge **before**
+         * calling this function
+         *
+         * **A Note on Counters:**<br>
+         * AppAttest only checks whether the signature counter **BEFORE SIGNING AN ASSERTION** is _higher_ than a provided value. That is, a value of `0` will always work.
+         * Warden Supreme also allows checking for a maximum, hence [validCounters]. If you don't care for it, just pass [Long.MAX_VALUE] as upper bound (inclusive).<br>
+         * **Also note that the upper bound will also be calculated on the value _BEFORE_ creating the assertion!**
+         *
+         * @param validatedAttestation the previously validated attestation
+         * @param expectedChallenge the expected client data to be contained in [assertion]
+         * @param validCounters The range of valid counters
+         */
+        fun verifyAssertion(
+            validatedAttestation: ValidatedAttestation,
+            assertion: ByteArray,
+            expectedChallenge: ByteArray,
+            validCounters: LongRange
+        ): Result<Assertion>
+
+        /**
+         * Verifies a fresh assertion, tied to a previously stored attestation.
+         *
+         * **The attestation is not verified again!**
+         * There is no timeliness guarantee of any kind, so manually verify the freshness of the challenge **before**
+         * calling this function
+         *
+         * **A Note on Counters:**<br>
+         * AppAttest only checks whether the signature counter is _higher_ than a provided value. That is, a value of `0` will always work.
+         * Warden Supreme also allows checking for a maximum, hence [validCounters]. If you don't care for it, just pass [Long.MAX_VALUE] as upper bound (inclusive).<br>
+         * **Also note that the upper bound will also be calculated on the value _BEFORE_ creating the assertion!**
+         *
+         *
+         * @param validatedAttestation the previously validated attestation
+         * @param validCounters The range of valid counters
+         * @param validator a fresh [AssertionChallengeValidator] that checks for challenge, clientData, etc.
+         */
+        fun verifyAssertion(
+            validatedAttestation: ValidatedAttestation,
+            assertion: ByteArray,
+            referenceClientData: ByteArray,
+            validCounters: LongRange,
+            expectedChallenge: ByteArray,
+            validator: AssertionChallengeValidator,
+        ): Result<Assertion>
     }
 
     /**
@@ -261,7 +331,7 @@ abstract class AttestationService {
 value class AssertionData private constructor(private val pair: Pair<ByteArray, ByteArray>) {
 
     /**
-     * Pairs an Apple AppAttest  assertion with the referenced clientData value
+     * Pairs an Apple AppAttest assertion with the referenced clientData value
      */
     constructor(assertion: ByteArray, clientData: ByteArray) : this(assertion to clientData)
 
@@ -332,18 +402,53 @@ object NoopAttestationService : AttestationService() {
                 runBlocking { verifyAttestation(listOf(attestationObject), challenge, clientData = null) }
 
             @DisabledAttestation
-            override fun verifyAssertion(
+            override fun verifyCombined(
                 attestationObject: ByteArray,
                 assertionFromDevice: ByteArray,
                 referenceClientData: ByteArray,
                 challenge: ByteArray,
-                counter: Long
             ) = runBlocking {
                 verifyAttestation(
                     listOf(attestationObject, assertionFromDevice),
                     challenge,
                     referenceClientData
                 )
+            }
+
+            //Will always throw
+            @DisabledAttestation
+            override fun verifyAssertion(
+                validatedAttestation: ValidatedAttestation,
+                assertion: ByteArray,
+                expectedChallenge: ByteArray,
+                validCounters: LongRange,
+            ): Result<Assertion> = verifyAssertion(
+                validatedAttestation, assertion, expectedChallenge, validCounters, expectedChallenge,
+                object : AssertionChallengeValidator {
+                    override fun validate(
+                        assertionObj: Assertion,
+                        clientData: ByteArray,
+                        attestationPublicKey: ECPublicKey,
+                        challenge: ByteArray
+                    ) = true
+                })
+
+
+            //Will always throw
+            @DisabledAttestation
+            override fun verifyAssertion(
+                validatedAttestation: ValidatedAttestation,
+                assertion: ByteArray,
+                referenceClientData: ByteArray,
+                validCounters: LongRange,
+                expectedChallenge: ByteArray,
+                validator: AssertionChallengeValidator,
+            ): Result<Assertion> {
+                val envelope = assertionReader.readValue<AssertionEnvelope>(assertion)
+                return catchingUnwrapped {
+                    Assertion(envelope.signature, envelope.authenticatorDataParsed())
+                }
+
             }
 
         }