+++ CHANGELOG.md
@@ -5,8 +5,13 @@ dedicated artefacts,
 this changelog also includes the original WARDEN changelog.
 
 # 1.0.0-SNAPSHOT
+* Features:
+    * Proper iOS assertion validation (thanks @andreybogdanov-sprind!)
+    * Serializers and ASN.1 codec for `ValidatedAttestation`
 * Fixes
     * Expose Kotlin-Stdlib as API dependency to make integration with Java projects smoother
+* API Changes:
+    * rename `AttestationService.ios.verifyAssertion` -> `AttestationService.ios.verifyCombined` and deprecate the old function 
 * Dependency Updates
     * Update upstream attestation library to 9066c0a003225e776b93ba5906d46c45904173de